use shared test fixture for random project
