+++ tests/convert_grid2d_test.py
@@ -1,6 +1,7 @@
 """Tests for PointSet validation"""
 
 import logging
+from pathlib import Path
 
 import numpy as np
 from geoh5py.workspace import Workspace
@@ -8,7 +9,7 @@ from geoh5py.workspace import Workspace
 import omf
 
 
-def test_grid2d_to_geoh5(tmp_path, caplog):
+def test_grid2d_to_geoh5(tmp_path: Path, caplog):
     """Test pointset geometry validation"""
 
     dip = np.random.uniform(low=0.0, high=90, size=1)
