+++ tests/convert_project_test.py
@@ -2,43 +2,38 @@
 
 import logging
 import os
+from pathlib import Path
 
 from geoh5py.workspace import Workspace
 
 import omf
 
-from .doc_example_test import TestDocEx
 
-
-def test_project_to_geoh5(tmp_path, caplog):
+def test_project_to_geoh5(random_project: omf.Project, tmp_path: Path, caplog):
     """Test pointset geometry validation"""
-    proj = TestDocEx.make_random_project()
-
     file = str(tmp_path / "project.geoh5")
 
-    omf.OMFWriter(proj, file)
+    omf.OMFWriter(random_project, file)
     warning_records = [r for r in caplog.records if r.levelno == logging.WARNING]
     assert len(warning_records) == 1
 
     with Workspace(file) as workspace:
-        assert len(workspace.objects) == len(proj.elements)
+        assert len(workspace.objects) == len(random_project.elements)
 
     project = omf.fileio.geoh5.GeoH5Reader(file).project
 
-    assert len(project.elements) == len(proj.elements)
+    assert len(project.elements) == len(random_project.elements)
 
 
-def test_project_compression(tmp_path):
+def test_project_compression(random_project: omf.Project, tmp_path: Path):
     """Test pointset geometry validation"""
-    proj = TestDocEx.make_random_project()
-
     file_low_comp = str(tmp_path / "project_low_comp.geoh5")
     file_med_comp = str(tmp_path / "project_med_comp.geoh5")
     file_high_comp = str(tmp_path / "project_high_comp.geoh5")
 
-    omf.OMFWriter(proj, file_low_comp, compression=1)
-    omf.OMFWriter(proj, file_med_comp, compression=5)
-    omf.OMFWriter(proj, file_high_comp, compression=9)
+    omf.OMFWriter(random_project, file_low_comp, compression=1)
+    omf.OMFWriter(random_project, file_med_comp, compression=5)
+    omf.OMFWriter(random_project, file_high_comp, compression=9)
 
     size_low_comp = os.stat(file_low_comp).st_size
     size_med_comp = os.stat(file_med_comp).st_size
