+++ tests/convert_surface_test.py
@@ -1,12 +1,14 @@
 """Tests for PointSet validation"""
 
+from pathlib import Path
+
 import numpy as np
 from geoh5py.workspace import Workspace
 
 import omf
 
 
-def test_surface_to_geoh5(tmp_path):
+def test_surface_to_geoh5(tmp_path: Path):
     """Test pointset geometry validation"""
     surf = omf.SurfaceElement(
         name="trisurf",
